✨ Add `time-since` fields for `get` operations
there is currently no easy way to check the amount of units of time
since something was created or updated in Github actions. This adds
a basic mechanism that parses the date time of the operation, and
produces new outputs that include the `time-since` the operation in
various different units that can be compared against.